Replace outdent with our own implementation
Currently we use the `outdent` package so that tests that involve
writing multi-line strings to files look a bit nicer. This package
doesn't play nice in an ESM context, and we actually already have
equivalents in `normalizeWhitespaceString` and `buildChangelog` that we
can use instead. So this commit removes `outdent` in favor of these
helpers.
