+++ package.json
@@ -61,7 +61,6 @@
     "jest-it-up": "^2.0.2",
     "jest-when": "^3.5.2",
     "nanoid": "^3.3.4",
-    "outdent": "^0.8.0",
     "prettier": "^2.2.1",
     "prettier-plugin-packagejson": "^2.3.0",
     "rimraf": "^4.0.5",
