+++ tests/helpers.ts
@@ -60,7 +60,7 @@ export async function withSandbox(fn: (sandbox: Sandbox) => any) {
   try {
     await fn({ directoryPath });
   } finally {
-    await rimraf(directoryPath);
+    await fs.promises.rm(directoryPath, { force: true, recursive: true });
   }
 }
 
@@ -91,4 +91,47 @@ export function isExecaError(error: unknown): error is ExecaError {
     hasProperty(error, 'isCanceled') &&
     hasProperty(error, 'exitCode')
   );
+}
+
+/**
+ * Given a string, resets its indentation and removes leading and trailing
+ * whitespace (except for a trailing newline).
+ *
+ * @param string - The string.
+ * @returns The normalized string.
+ */
+export function normalizeMultilineString(string: string): string {
+  const lines = string
+    .replace(/^[\n\r]+/u, '')
+    .replace(/[\n\r]+$/u, '')
+    .split('\n');
+  const indentation = lines[0].match(/^([ ]+)/u)?.[1] ?? '';
+  const normalizedString = lines
+    .map((line) => {
+      return line.replace(new RegExp(`^${indentation}`, 'u'), '');
+    })
+    .join('\n')
+    .trim();
+  return `${normalizedString}\n`;
+}
+
+/**
+ * Builds a changelog by filling in the first part automatically, which never
+ * changes.
+ *
+ * @param variantContent - The part of the changelog that can change depending
+ * on what is expected or what sort of changes have been made to the repo so
+ * far.
+ * @returns The full changelog.
+ */
+export function buildChangelog(variantContent: string): string {
+  const invariantContent = normalizeMultilineString(`
+    # Changelog
+    All notable changes to this project will be documented in this file.
+
+    The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+    and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+  `);
+
+  return `${invariantContent}\n${normalizeMultilineString(variantContent)}`;
 }
